fix: address PR #10 review — command injection, SSRF, auth hardening, CI
Rework the still-valid code/CI findings from the wrokin review and CodeQL:

- node github.executor: replace shell exec with execFile + arg arrays;
  validate repository/branch/taskId (no flags, no traversal) before use
  (CodeQL uncontrolled-command-line, branch flag-injection)
- gateway proxy: pin forwarded URL to the trusted service origin, reject
  drift (CodeQL SSRF)
- registration-token guard: constant-time comparison via timingSafeEqual
- api-key guard: fail closed (throw) when API_KEY unset in production
- auth/app modules: drop hardcoded JWT fallback; getJwtSecret() fails fast
- shared: add config/env helpers; harden REPOSITORY/BRANCH patterns +
  isValidRepository/isValidBranch; allow fractional CPU in heartbeat schema
- CI: add least-privilege permissions block; gate npm audit at critical,
  report high non-blocking
- docs: fix mixed-language artifact in community.md; rename mobile.md ->
  tooling.md and bizdev.md -> community-devrel.md to match rewritten roles

+++ packages/shared/src/schemas/heartbeat.schema.ts
@@ -5,8 +5,9 @@ export const heartbeatRequestSchema = z.object({
 	status: z.enum(["online", "offline", "busy"]),
 	resources: z.object({
 		cpu: z.object({
-			used: z.number().int().min(0),
-			total: z.number().int().min(1),
+			// Fractional core usage is normal (e.g. 0.5 cores), so don't force ints.
+			used: z.number().min(0),
+			total: z.number().min(1),
 		}),
 		ram: z.object({
 			used: z.string().min(1),
